src/requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
bson==0.5.8
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.1.2
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
djongo==1.3.6
factory-boy==3.2.0
Faker==8.2.1
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==20.9
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.5
